+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to the "zowe-explorer-api" extension will be documented in this file.
 
-## `2.4.1`
+## TBD Release
+
+### New features and enhancements
+
+- New API call `getJobsByParameters` to allow filtering jobs by status.
+
+## 2.4.1
 
 ### Bug fixes